Added Composable interface, refactored constructor to work with characteristics with no starting data.
